+++ schemas/Role.js
@@ -3,5 +3,5 @@
 
   name: { type: 'string', unique: true },
   active: { type: 'boolean', default: true },
-  unit: 'Unit',
+  unit: { type: 'Unit', delete: 'restrict' },
 });